Rename tim_generator and tim_translator, update imports
